requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
httpx
cryptography